移动数据结构体到IO基类
