+++ MU3Input/IO/HidIO.cs
@@ -6,40 +6,6 @@ using SimpleHID.Raw;
 
 namespace MU3Input
 {
-    [StructLayout(LayoutKind.Sequential, Pack = 1, Size = 64)]
-    public struct OutputData
-    {
-        [MarshalAs(UnmanagedType.ByValArray, SizeConst = 10, ArraySubType = UnmanagedType.U1)]
-        public byte[] Buttons;
-
-        public ushort Lever;
-        [MarshalAs(UnmanagedType.U1)] public bool Scan;
-
-        [MarshalAs(UnmanagedType.ByValArray, SizeConst = 10)]
-        public byte[] AimiId;
-    }
-
-    [StructLayout(LayoutKind.Sequential, Pack = 1, Size = 64)]
-    public unsafe struct SetLedInput
-    {
-        public byte Type;
-        public byte LedBrightness;
-
-        [MarshalAs(UnmanagedType.ByValArray, SizeConst = 10)]
-        public fixed byte LedColors[3 * 10];
-    }
-
-    [StructLayout(LayoutKind.Sequential, Pack = 1, Size = 64)]
-    public unsafe struct SetOptionInput
-    {
-        public byte Type;
-
-        [MarshalAs(UnmanagedType.ByValArray, SizeConst = 10)]
-        public fixed byte AimiId[10];
-    }
-
-    
-
     // ReSharper disable once InconsistentNaming
     public class HidIO : IO
     {
